% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_names.R
\name{clean_names}
\alias{clean_names}
\title{Cleans species names with redundant white spaces and character case differences}
\usage{
clean_names(species)
}
\arguments{
\item{species}{Species name to be cleaned}
}
\value{
Cleaned species name
}
\description{
Cleans species names with redundant white spaces and character case differences
}
\examples{

sp <- clean_names(species='Lates Niloticus')

}
\seealso{
{\code{\link{priorityindex}}, \code{\link{clean_names}}}
}
