% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpi.R
\name{cpi}
\alias{cpi}
\title{Conservation}
\usage{
cpi(data, habitat, species, area, iucn, plot = NULL)
}
\arguments{
\item{data}{A dataframe with all the details, of which the habitat names, surface, species names and threat status should be provided.}

\item{habitat}{A column with the different habitat names considered.}

\item{species}{A column with all species names in the habitat.}

\item{area}{The habitat area or measure of coverage that can used for conservation selection.}

\item{iucn}{A column with the speciea threat classes or category such as vulnerable.}

\item{plot}{To visualise the a bar graph of priority score for each habitat.}
}
\value{
Dataframe with species scores
}
\description{
Conservation
}
\examples{

victoria <- cpi(data=gbif, habitat='habitat', species='species', area='area', iucn='iucn')

}
