% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello world for saying Hello}
\usage{
hello(x)
}
\arguments{
\item{x}{The number to be added to}
}
\description{
Hello world for saying Hello
}
\examples{
hello(7)
\dontrun{
hello(9)
}
}
