% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorityindex.R
\name{priorityindex}
\alias{priorityindex}
\title{Conservation priority score for a particular habitat}
\usage{
priorityindex(
  data,
  habitat,
  species,
  area,
  iucn,
  lat = NULL,
  lon = NULL,
  polygon = NULL,
  select = NULL,
  subset = NULL,
  plot = FALSE,
  full = FALSE,
  map = FALSE
)
}
\arguments{
\item{data}{Data frame with the habitat details including habitat, species, area and IUCN status for
for each species}

\item{habitat}{A variable for habitat names. These can be water bodies, or unit were conservation priorities
can be conducted in comparison to the other. For example, habitats can be lakes or the different habitat
groups such as offshore, inshore or middle sections of the river. For rivers, habitats can be subcatchments of
river segments or individuals river compared. The surface area or any unit used to measure the
extend of the habitat should be particular to it and not generalaised across the habitats.}

\item{species}{A variable for species names, whether scientific or local names.}

\item{area}{A variable with the area of the habitat. Either measured in the field or using already made polygons.}

\item{iucn}{A variable for species IUCN categories based on IUCN RedList (IUCN 2001).
The package is currently tailored to IUCN Red List status assessments.}

\item{select}{Particular habitat name to compute the priority scores for.}

\item{plot}{To visualize the a bar graph of priority score for each habitat.}
}
\value{
Conservation priority score for one habitat form the data.
}
\description{
Conservation priority score for a particular habitat
}
\details{
For all species were weighted based on the IUCN Red List status.
Highest and lowest weights assigned to extinct and least concern conservation categories, respectively.
The weights were assigned as follows: \strong{ET = 7}, \strong{EXw = 6}, \strong{CR = 5},
\strong{DD = 5}, \strong{NE = 5}, \strong{EN = 4}, \strong{VU = 3}, \strong{NT = 2},
and \strong{LC = 1}. The conservation priority scores are computed in \strong{(Basooma et al., 2022)}
}
\examples{

\dontrun{
#species record id
id <- seq(1, 30, 1)
#lakes to be prioritized for conservation
habitats <- rep(c('Kyoga','Victoria','Albert'), 10)
#surface area for each lake
surface_area <- rep(c(1821.6, 33700, 2850), 10) #Uganda surface area in Uganda

#species recorded in each lake
species <- c(rep('Haplochromine latifasciatus', 2),
            rep('Lates macropthalmous', 1),
            rep('Haplochromis phytophagus',1),
            rep('L. niloticus',21),
            rep('Clarias gariepinus', 5))
#final dataframe
df_final <- data.frame(id, habitats, surface_area, species)
#Assign each species the IUCN categories based on IUCN RedList
df_final$iucn <- ifelse(species=='Clarias gariepinus', 'LC',
                        ifelse(species=='L. niloticus', 'LC',
                              ifelse(species=='Haplochromis phytophagus', 'DD',
                                    ifelse(species=='Lates macropthalmous','EN',
                                           ifelse(species=='Haplochromine latifasciatus',
                                           'CR', NA)))))

victoriacpi <- cpi_one(data= df_final, habitat='habitats', species='species',
area='surface_area',iucn='iucn', select='Victoria')

library(sf)

data('gbif')

gbif

victoria <- cpi_one(data=gbif, habitat='waterbody', species='species', area='surfacearea',
iucn='iucnstatus', select='Lake Victoria')
}


}
\references{
\enumerate{
\item Basooma, A., Nakiyende, H., Olokotum, M., Balirwa, J. S., Nkalubo, W.,
Musinguzi, L., & Natugonza, V. (2022). A novel index to aid in prioritizing habitats
for site‐based conservation. Ecology and Evolution, 12(3), e8762.
\item Natural Resources. Species Survival Commission, & IUCN Species Survival Commission. (2001).
IUCN Red List categories and criteria. IUCN.
}
}
\seealso{
{\code{\link{clean_names}}}
}
\author{
Anthony Basooma (bas4ster@gmail.com)
}
