% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarendm.R
\name{rareend}
\alias{rareend}
\title{Endemicty and rarity}
\usage{
rareend(data, habitat, species, area, iucn, hname)
}
\arguments{
\item{data}{fhfh}

\item{habitat}{fhfh}

\item{species}{jfjf}

\item{area}{jfjfj}

\item{iucn}{jfjf}

\item{hname}{nnvnv}
}
\value{
priority score for one habitat form the dataset.
}
\description{
Endemicty and rarity
}
\examples{

\dontrun{
library(sf)

data('gbif')

gbif

victoria <- cpi(data=gbif, habitat='wb',
species='species', area='sa', iucn='IUCN', hname='Lake Victoria')
}



}
