% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarity.R
\name{rarity}
\alias{rarity}
\title{Computes the species endemicity weights based on the number habitats is recorded in.}
\usage{
rarity(data, habitat, species, sp)
}
\arguments{
\item{data}{Data frame with all species and habitats names in the area of study.}

\item{habitat}{A column with the habitat names.}

\item{species}{A column with species names.}

\item{sp}{The species names in question. Either scientific or local but must be in the data.}
}
\value{
value, rarity or endemic score.
}
\description{
Computes the species endemicity weights based on the number habitats is recorded in.
}
\examples{

\dontrun{

library(sf)

data('gbif')

gbif

rarity_lates <- rarity(data= gbif, habitat='wb',
species='species', sp='Lates niloticus')

}


}
