% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_iucn.R
\name{clean_iucn}
\alias{clean_iucn}
\title{Checks and standardizes species threat status .}
\usage{
clean_iucn(iucn)
}
\arguments{
\item{iucn}{The species International Union for C.}
}
\value{
string
}
\description{
Checks and standardizes species threat status .
}
\examples{

sp <- clean_iucn(iucn='cr')

}
