% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_names.R
\name{clean_names}
\alias{clean_names}
\title{Checks the species names to match databses such as FishBase}
\usage{
clean_names(species)
}
\arguments{
\item{species}{Species name to be cleaned}
}
\value{
string
}
\description{
Checks the species names to match databses such as FishBase
}
\examples{

sp <- clean_names(species='Lates Niloticus')

}
